refactor: Gamer -> Participant 클래스명 변경
- 게임 참가자로서의 성격이 드러나도록 클래스명을 수정한다.
