+++ src/main/java/blackjack/domain/participant/GameRule.java
@@ -1,7 +1,7 @@
 package blackjack.domain.participant;
 
 import blackjack.domain.card.Hand;
-import blackjack.domain.participant.gamer.Player;
+import blackjack.domain.participant.participant.Player;
 import blackjack.domain.result.ProfitResult;
 import java.util.Map;
