+++ src/main/java/blackjack/domain/participant/Participants.java
@@ -1,7 +1,7 @@
 package blackjack.domain.participant;
 
 import blackjack.domain.card.Hand;
-import blackjack.domain.participant.gamer.Dealer;
+import blackjack.domain.participant.participant.Dealer;
 import blackjack.domain.result.ProfitResult;
 import java.util.List;
 import java.util.Map;
