+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -15,8 +15,8 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.CardScore;
 import blackjack.domain.card.Hand;
 import blackjack.domain.card.Shape;
-import blackjack.domain.participant.gamer.Dealer;
-import blackjack.domain.participant.gamer.Player;
+import blackjack.domain.participant.participant.Dealer;
+import blackjack.domain.participant.participant.Player;
 import blackjack.domain.result.ProfitResult;
 import java.util.List;
 import java.util.Map;
