+++ src/test/java/blackjack/domain/participant/PlayersTest.java
@@ -7,7 +7,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import blackjack.domain.card.Hand;
-import blackjack.domain.participant.gamer.Player;
+import blackjack.domain.participant.participant.Player;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
